fix: fix menu drawer logic
Fixed the menu drawer and spun the pop-up into its own file.
